refactored out code into more helper files
