+++ acceptance_tests/features/steps/print_file.py
@@ -1,6 +1,3 @@
-import copy
-import hashlib
-import json
 import logging
 
 from behave import then, step
@@ -8,8 +5,8 @@ from google.cloud import storage
 from retrying import retry
 from structlog import wrap_logger
 
-from acceptance_tests.utilities.mappings import PACK_CODE_TO_SFTP_DIRECTORY, PACK_CODE_TO_DATASET, \
-    PACK_CODE_TO_DESCRIPTION
+from acceptance_tests.utilities.manifest_file_helper import check_manifest_files_created
+from acceptance_tests.utilities.mappings import PACK_CODE_TO_SFTP_DIRECTORY
 from acceptance_tests.utilities.print_file_helper import \
     create_expected_csv_lines, create_expected_on_request_questionnaire_csv, \
     create_expected_supplementary_materials_csv, create_expected_reminder_letter_csv_lines, \
@@ -17,35 +14,24 @@ from acceptance_tests.utilities.print_file_helper import \
     create_expected_csv_lines_for_ce_estab_responses, create_expected_CE_Estab_questionnaire_csv_lines, \
     create_expected_questionnaire_csv_lines, create_expected_Welsh_CE_Estab_questionnaire_csv_line_endings, \
     create_expected_HH_UAC_supplementary_materials_csv, \
-    create_expected_csv_lines_for_reminder_survey_launched
+    create_expected_csv_lines_for_reminder_survey_launched, check_print_files_have_all_the_expected_data
 from acceptance_tests.utilities.sftp_utility import SftpUtility
 from acceptance_tests.utilities.test_case_helper import test_helper
 from config import Config
 
 logger = wrap_logger(logging.getLogger(__name__))
 
-ICL_PACKCODES_WHICH_ARE_SORTED = [
-    'D_CE1A_ICLCR1', 'D_CE1A_ICLCR2B', 'D_ICA_ICLR1', 'D_ICA_ICLR1',
-    'D_ICA_ICLR2B', 'D_ICA_ICLR2B', 'D_CE4A_ICLR4', 'D_CE4A_ICLS4'
-]
-ICL_TEMPLATE_FIELD_OFFICER_COLUMN = 14
-ICL_TEMPLATE_ORGANISATION_COLUMN = 12
-
-QM_PACKCODES_WHICH_ARE_SORTED = ['D_FDCE_I1', 'D_FDCE_I2', 'D_FDCE_I4', 'D_FDCE_H1', 'D_FDCE_H2']
-QM_TEMPLATE_FIELD_OFFICER_COLUMN = 15
-QM_TEMPLATE_ORGANISATION_COLUMN = 14
-
 
 @then('correctly formatted "{pack_code}" print files are created for questionnaire')
 def check_correct_questionnaire_files_on_sftp_server(context, pack_code):
     expected_csv_lines = create_expected_questionnaire_csv_lines(context, pack_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
 
 
 @then('correctly formatted "{pack_code}" print files are created for CE Estab questionnaires')
 def check_correct_CE_Estab_questionnaire_files_on_sftp_server(context, pack_code):
     expected_csv_lines = create_expected_CE_Estab_questionnaire_csv_lines(context, pack_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
 
 
 @then('correctly formatted "{pack_code}" print files are created for CE Estab Welsh questionnaires')
@@ -118,48 +104,48 @@ def check_ce_estab_welsh_questionnaire_is_correct(context, pack_code, expected_p
 def check_correct_files_on_sftp_server(context, pack_code):
     context.expected_pack_code = pack_code
     expected_csv_lines = create_expected_csv_lines(context, pack_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
 
 
 @then('correctly formatted "{pack_code}" print files are created for CE Estab expected responses')
 def check_correct_ce_estab_files_on_sftp_server(context, pack_code):
     context.expected_pack_code = pack_code
     expected_csv_lines = create_expected_csv_lines_for_ce_estab_responses(context, pack_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
 
 
 @then('correctly formatted "{pack_code}" reminder letter print files are created')
 def check_correct_reminder_letter_files_on_sftp_server(context, pack_code):
     expected_csv_lines = create_expected_reminder_letter_csv_lines(context, pack_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
 
 
 @then('correctly formatted "{}" reminder letter print files are created for the survey launched cases')
 def check_correct_reminder_letter_files_for_survey_launched_on_sftp_server(context, pack_code):
     expected_csv_lines = create_expected_reminder_letter_csv_lines(context, pack_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
 
 
 @then('correctly formatted "{pack_code}" reminder questionnaire print files are created')
 def check_correct_reminder_questionnaire_files_on_sftp_server(context, pack_code):
     expected_csv_lines = create_expected_reminder_questionnaire_csv_lines(context, pack_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
 
 
 @step('the refused case of type "{refusal_type}" only appears in the "{pack_code}" print files if it is a HARD_REFUSAL')
-def check_correct_unrefused_files_on_sftp_server(context, pack_code, refusal_type):
+def check_correct_unrefused_files_on_sftp_server(context, refusal_type, pack_code):
     if refusal_type == 'EXTRAORDINARY_REFUSAL':
         expected_csv_lines = create_expected_csv_lines(context, pack_code, context.refused_case_id)
     else:
         expected_csv_lines = create_expected_csv_lines(context, pack_code)
 
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
 
 
 @then('only unreceipted cases appear in "{pack_code}" print files')
 def check_correct_unreceipted_files_on_sftp_server(context, pack_code):
     expected_csv_lines = create_expected_csv_lines(context, pack_code, context.first_case['id'])
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
 
 
 @step('skeleton cases do not appear in "{pack_code}" print files')
@@ -171,7 +157,7 @@ def check_skeleton_case_not_in_print_file(context, pack_code):
 @then('there is a correct "{pack_code}" manifest file for each csv file written')
 def check_manifest_files(context, pack_code):
     logger.debug("checking manifest files exist for csv files")
-    _check_manifest_files_created(context, pack_code)
+    check_manifest_files_created(context, pack_code)
 
 
 @step('correctly formatted on request contn questionnaire print and manifest files for "{fulfilment_code}" are created')
@@ -179,8 +165,8 @@ def check_manifest_files(context, pack_code):
 @step('correctly formatted on request UAC questionnaire print and manifest files for "{fulfilment_code}" are created')
 def correct_on_request_questionnaire_print_files(context, fulfilment_code):
     expected_csv_lines = create_expected_on_request_questionnaire_csv(context, fulfilment_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, fulfilment_code)
-    _check_manifest_files_created(context, fulfilment_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, fulfilment_code)
+    check_manifest_files_created(context, fulfilment_code)
 
 
 @step(
@@ -188,30 +174,24 @@ def correct_on_request_questionnaire_print_files(context, fulfilment_code):
     'print and manifest files for "{fulfilment_code}" are created')
 def correct_on_request_fulfilment_questionnaire_print_files(context, fulfilment_code):
     expected_csv_lines = create_expected_on_request_fulfilment_questionnaire_csv(context, fulfilment_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, fulfilment_code)
-    _check_manifest_files_created(context, fulfilment_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, fulfilment_code)
+    check_manifest_files_created(context, fulfilment_code)
 
 
 @then('correctly formatted on request supplementary material print'
       ' and manifest files for "{fulfilment_code}" are created')
 def correct_supplementary_material_print_files(context, fulfilment_code):
     expected_csv_lines = create_expected_supplementary_materials_csv(context, fulfilment_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, fulfilment_code)
-    _check_manifest_files_created(context, fulfilment_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, fulfilment_code)
+    check_manifest_files_created(context, fulfilment_code)
 
 
 @step('correctly formatted on request HH UAC supplementary material print'
       ' and manifest files for "{fulfilment_code}" are created')
 def correct_HH_UAC_supplementary_material_print_files(context, fulfilment_code):
     expected_csv_lines = create_expected_HH_UAC_supplementary_materials_csv(context, fulfilment_code)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, fulfilment_code)
-    _check_manifest_files_created(context, fulfilment_code)
-
-
-def _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code):
-    with SftpUtility() as sftp_utility:
-        _validate_print_file_content(context, sftp_utility, context.test_start_local_datetime, expected_csv_lines,
-                                     pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, fulfilment_code)
+    check_manifest_files_created(context, fulfilment_code)
 
 
 def _get_print_file_rows_as_list(context, pack_code):
@@ -219,56 +199,6 @@ def _get_print_file_rows_as_list(context, pack_code):
         context.expected_print_files = sftp_utility.get_all_files_after_time(context.test_start_local_datetime,
                                                                              pack_code, ".csv.gpg")
         return sftp_utility.get_files_content_as_list(context.expected_print_files, pack_code)
-
-
-def check_actual_file_contents_sorted_by_production_code(unsorted_actual_content_list, pack_code):
-    # If this was split over multiple files it could fail, not expected with current testing
-    # produce a list of split_csv rows to sort
-    split_csv_rows = [
-        csvrow.split("|")
-        for csvrow in unsorted_actual_content_list
-    ]
-
-    # If called with an invalid packcode the test will fail, this is expected behaviour
-    sorted_list = []
-
-    # This will sort a list of lists based on template ICL or QM
-    # in both cases we're sorting by field_officer_id and org_name
-    if pack_code in ICL_PACKCODES_WHICH_ARE_SORTED:
-        sorted_list = sorted(split_csv_rows, key=lambda row: (row[ICL_TEMPLATE_FIELD_OFFICER_COLUMN],
-                                                              row[ICL_TEMPLATE_ORGANISATION_COLUMN]))
-    elif pack_code in QM_PACKCODES_WHICH_ARE_SORTED:
-        sorted_list = sorted(split_csv_rows, key=lambda row: (row[QM_TEMPLATE_FIELD_OFFICER_COLUMN],
-                                                              row[QM_TEMPLATE_ORGANISATION_COLUMN]))
-
-    # Turn back to a list of expected_csv rows
-    expected_csv_rows = [
-        '|'.join(row)
-        for row in sorted_list
-    ]
-
-    test_helper.assertEquals(unsorted_actual_content_list, expected_csv_rows,
-                             'Sorted file contents did not match expected')
-
-
-@retry(retry_on_exception=lambda e: isinstance(e, FileNotFoundError), wait_fixed=1000, stop_max_attempt_number=120)
-def _validate_print_file_content(context, sftp_utility, start_of_test, expected_csv_lines, pack_code):
-    logger.debug('Checking for files on SFTP server')
-
-    context.expected_print_files = sftp_utility.get_all_files_after_time(start_of_test, pack_code, ".csv.gpg")
-
-    actual_content_list = sftp_utility.get_files_content_as_list(context.expected_print_files, pack_code)
-    if not actual_content_list:
-        raise FileNotFoundError
-
-    unsorted_actual_content_list = copy.deepcopy(actual_content_list)
-    actual_content_list.sort()
-    expected_csv_lines.sort()
-
-    test_helper.assertEquals(actual_content_list, expected_csv_lines, 'Print file contents did not match expected')
-
-    if pack_code in ICL_PACKCODES_WHICH_ARE_SORTED or pack_code in QM_PACKCODES_WHICH_ARE_SORTED:
-        check_actual_file_contents_sorted_by_production_code(unsorted_actual_content_list, pack_code)
 
 
 @retry(retry_on_exception=lambda e: isinstance(e, FileNotFoundError), wait_fixed=1000, stop_max_attempt_number=120)
@@ -280,69 +210,6 @@ def _check_print_file_does_not_contain_case_ref(context, start_of_test, sftp_uti
         raise FileNotFoundError
     test_helper.assertFalse(any(
         context.case_created_events[0]['payload']['collectionCase']['caseRef'] in row for row in actual_content_list))
-
-
-def _check_manifest_files_created(context, pack_code):
-    with SftpUtility() as sftp_utility:
-        files = sftp_utility.get_all_files_after_time(context.test_start_local_datetime, pack_code)
-
-        for _file in files:
-            if _file.filename.endswith(".csv.gpg"):
-                csv_file = _file
-                manifest_file = _get_matching_manifest_file(csv_file.filename, files)
-
-                if manifest_file is None:
-                    test_helper.fail(f'Failed to find manifest file for {csv_file.filename}')
-
-                actual_manifest = _get_actual_manifest(sftp_utility, manifest_file, pack_code)
-                creation_datetime = actual_manifest['manifestCreated']
-                expected_manifest = _create_expected_manifest(sftp_utility, csv_file, creation_datetime, pack_code)
-                test_helper.assertDictEqual(actual_manifest, expected_manifest)
-
-
-def _get_actual_manifest(sftp_utility, manifest_file, pack_code):
-    actual_manifest_json = sftp_utility.get_file_contents_as_string(f'{PACK_CODE_TO_SFTP_DIRECTORY[pack_code]}/'
-                                                                    f'{manifest_file.filename}')
-    return json.loads(actual_manifest_json)
-
-
-def _get_matching_manifest_file(filename, files):
-    manifest_filename = filename.replace(".csv.gpg", ".manifest")
-
-    for _file in files:
-        if _file.filename == manifest_filename:
-            return _file
-
-    return None
-
-
-def _create_expected_manifest(sftp_utility, csv_file, created_datetime, pack_code):
-    actual_file_contents = sftp_utility.get_file_contents_as_string(f'{PACK_CODE_TO_SFTP_DIRECTORY[pack_code]}'
-                                                                    f'/{csv_file.filename}')
-    decrypted_file_contents = sftp_utility.decrypt_message(actual_file_contents)
-
-    md5_hash = hashlib.md5(actual_file_contents.encode('utf-8')).hexdigest()
-    expected_size = sftp_utility.get_file_size(f'{PACK_CODE_TO_SFTP_DIRECTORY[pack_code]}/{csv_file.filename}')
-
-    _file = dict(
-        sizeBytes=str(expected_size),
-        md5sum=md5_hash,
-        relativePath='./',
-        name=csv_file.filename,
-        rows=len(decrypted_file_contents.splitlines())
-    )
-
-    manifest = dict(
-        schemaVersion='1',
-        files=[_file],
-        sourceName="ONS_RM",
-        manifestCreated=created_datetime,
-        description=PACK_CODE_TO_DESCRIPTION[pack_code],
-        dataset=PACK_CODE_TO_DATASET[pack_code],
-        version='1'
-    )
-
-    return manifest
 
 
 @step("the files have all been copied to the bucket")
@@ -375,4 +242,4 @@ def compare_sftp_with_gcp_files(context, bucket, filename):
 def check_reminder_files_with_survey_launched(context, pack_code):
     expected_csv_lines = create_expected_csv_lines_for_reminder_survey_launched(context, pack_code,
                                                                                 context.survey_started_case_ids)
-    _check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
+    check_print_files_have_all_the_expected_data(context, expected_csv_lines, pack_code)
