+++ acceptance_tests/features/steps/receipt.py
@@ -3,18 +3,19 @@ import json
 
 from behave import step
 
-from acceptance_tests.features.steps.case_look_up import get_ccs_qid_for_case_id
-from acceptance_tests.features.steps.event_log import check_if_event_list_is_exact_match
-from acceptance_tests.features.steps.fulfilment import send_print_fulfilment_request
-from acceptance_tests.features.steps.telephone_capture import request_individual_telephone_capture, \
-    check_correct_uac_updated_message_is_emitted, request_hi_individual_telephone_capture
-from acceptance_tests.features.steps.unaddressed import send_questionnaire_linked_msg_to_rabbit, \
-    check_uac_message_is_received
+from acceptance_tests.utilities.case_api_helper import get_ccs_qid_for_case_id
+from acceptance_tests.utilities.event_helper import check_if_event_list_is_exact_match
+from acceptance_tests.utilities.fieldwork_helper import field_work_cancel_callback
+from acceptance_tests.utilities.fulfilment_helper import send_print_fulfilment_request
 from acceptance_tests.utilities.pubsub_helper import publish_to_pubsub
 from acceptance_tests.utilities.rabbit_context import RabbitContext
 from acceptance_tests.utilities.rabbit_helper import start_listening_to_rabbit_queue, store_all_msgs_in_context, \
     check_no_msgs_sent_to_queue
+from acceptance_tests.utilities.telephone_capture_helper import request_hi_individual_telephone_capture, \
+    request_individual_telephone_capture, check_correct_uac_updated_message_is_emitted
 from acceptance_tests.utilities.test_case_helper import test_helper
+from acceptance_tests.utilities.unadressed_helper import send_questionnaire_linked_msg_to_rabbit, \
+    check_uac_message_is_received
 from config import Config
 
 
@@ -103,7 +104,7 @@ def uac_updated_msg_emitted(context):
 def action_cancel_sent_to_fwm(context, address_type):
     context.messages_received = []
     start_listening_to_rabbit_queue(Config.RABBITMQ_OUTBOUND_FIELD_QUEUE, functools.partial(
-        _field_work_cancel_callback, context=context))
+        field_work_cancel_callback, context=context))
 
     test_helper.assertEqual(context.fwmt_emitted_case_id, context.first_case["id"])
     test_helper.assertEqual(context.addressType, address_type)
@@ -133,21 +134,6 @@ def case_updated_msg_sent_with_values(context, case_field, expected_field_value,
     else:
         test_helper.assertEqual(emitted_case['id'], context.receipting_case['id'])
         test_helper.assertEqual(str(emitted_case[case_field]), expected_field_value)
-
-
-def _field_work_cancel_callback(ch, method, _properties, body, context):
-    action_cancel = json.loads(body)
-
-    if not action_cancel['actionInstruction'] == 'CANCEL':
-        ch.basic_nack(delivery_tag=method.delivery_tag)
-        test_helper.fail(f'Unexpected message on {Config.RABBITMQ_OUTBOUND_FIELD_QUEUE} case queue. '
-                         f'Got "{action_cancel["actionInstruction"]}", wanted "CANCEL"')
-
-    context.addressType = action_cancel['addressType']
-    context.fwmt_emitted_case_id = action_cancel['caseId']
-    context.field_action_cancel_message = action_cancel
-    ch.basic_ack(delivery_tag=method.delivery_tag)
-    ch.stop_consuming()
 
 
 def _publish_object_finalize(context, case_id="0", tx_id="3d14675d-a25d-4672-a0fe-b960586653e8", questionnaire_id="0"):
