+++ acceptance_tests/features/steps/refusal.py
@@ -5,9 +5,8 @@ import requests
 from behave import step
 from retrying import retry
 
-from acceptance_tests.features.steps.event_log import check_if_event_list_is_exact_match
-from acceptance_tests.features.steps.field_reminder import fieldwork_message_callback
-from acceptance_tests.features.steps.receipt import _field_work_cancel_callback
+from acceptance_tests.utilities.event_helper import check_if_event_list_is_exact_match
+from acceptance_tests.utilities.fieldwork_helper import fieldwork_message_callback, field_work_cancel_callback
 from acceptance_tests.utilities.rabbit_context import RabbitContext
 from acceptance_tests.utilities.rabbit_helper import start_listening_to_rabbit_queue
 from acceptance_tests.utilities.test_case_helper import test_helper
@@ -88,7 +87,7 @@ def check_case_events(context):
 @step('a CANCEL action instruction is emitted to FWMT')
 def refusal_received(context):
     start_listening_to_rabbit_queue(Config.RABBITMQ_OUTBOUND_FIELD_QUEUE,
-                                    functools.partial(_field_work_cancel_callback, context=context))
+                                    functools.partial(field_work_cancel_callback, context=context))
 
     test_helper.assertEqual(context.fwmt_emitted_case_id, context.refused_case_id)
     test_helper.assertEqual(context.addressType, 'HH')
