+++ acceptance_tests/features/steps/sample.py
@@ -1,7 +1,6 @@
 from behave import step
 
-
-from acceptance_tests.features.steps.action_rules import poll_until_sample_is_ingested_to_action
+from acceptance_tests.utilities.action_helper import poll_until_sample_is_ingested_to_action
 from acceptance_tests.utilities.event_helper import get_and_check_case_created_messages, \
     get_and_check_uac_updated_messages, get_and_test_case_and_uac_msgs_are_correct
 from acceptance_tests.utilities.sample_load_helper import load_sample_file_helper
@@ -14,6 +13,7 @@ def load_sample_file_step(context, sample_file_name):
 
 @step('sample file "{sample_file_name}" is loaded successfully with census action plan collection ids')
 def load_sample_file_successfully_with_census_action_plan_collection_ids(context, sample_file_name):
+    # The actual values should be set with the annotation:  @hardcoded_census_values_for_collection_and_action_plan_ids
     load_sample_file_successfully_step(context, sample_file_name, check_loaded_into_db=False)
 
 
@@ -27,7 +27,9 @@ def load_sample_file_successfully_step(context, sample_file_name, check_loaded_i
     # TODO: in a nice world this would be well nicer
     # However when we set up the skeleton cases we need to use the census action-plan-id, which is fixed so the query
     # returns old matching rows too.
-    # So we then add in > start_of_test_datetime, excellent, except db is on UTC. So doesn't work :(
+    # So we then add in > start_of_test_datetime, excellent, except db is on UTC,
+    # sometimes different on local machine. So doesn't work
+    # could possibly get the DB start time at the start of the test, but a bit OTT
     if check_loaded_into_db:
         poll_until_sample_is_ingested_to_action(context)
