+++ acceptance_tests/features/steps/unaddressed.py
@@ -1,4 +1,3 @@
-import functools
 import json
 import logging
 import subprocess
@@ -13,8 +12,9 @@ from retrying import retry
 from structlog import wrap_logger
 
 from acceptance_tests.utilities.rabbit_context import RabbitContext
-from acceptance_tests.utilities.rabbit_helper import start_listening_to_rabbit_queue
 from acceptance_tests.utilities.test_case_helper import test_helper
+from acceptance_tests.utilities.unadressed_helper import send_questionnaire_linked_msg_to_rabbit, \
+    check_uac_message_is_received
 from config import Config
 
 logger = wrap_logger(logging.getLogger(__name__))
@@ -110,12 +110,8 @@ def check_blank_link_message_is_received(context):
 
 
 @step("a UACUpdated message not linked to a case is emitted to RH and Action Scheduler")
-def check_uac_message_is_received(context):
-    context.expected_message_received = False
-    start_listening_to_rabbit_queue(Config.RABBITMQ_RH_OUTBOUND_UAC_QUEUE,
-                                    functools.partial(_uac_callback, context=context))
-
-    test_helper.assertTrue(context.expected_message_received)
+def check_uac_message_is_received_step(context):
+    check_uac_message_is_received(context)
 
 
 @step("a Questionnaire Linked event is logged")
@@ -162,22 +158,6 @@ def get_case_id_by_questionnaire_id(questionnaire_id):
     test_helper.assertEqual(response.status_code, 200, "Unexpected status code")
     response_json = response.json()
     return response_json['id']
-
-
-def _uac_callback(ch, method, _properties, body, context):
-    parsed_body = json.loads(body)
-
-    if not parsed_body['event']['type'] == 'UAC_UPDATED':
-        ch.basic_nack(delivery_tag=method.delivery_tag)
-        return
-
-    test_helper.assertEqual(64, len(parsed_body['payload']['uac']['uacHash']))
-    context.expected_questionnaire_id = parsed_body['payload']['uac']['questionnaireId']
-    test_helper.assertEqual(context.expected_questionnaire_type, context.expected_questionnaire_id[:2])
-    context.expected_uac = parsed_body['payload']['uac']['uac']
-    context.expected_message_received = True
-    ch.basic_ack(delivery_tag=method.delivery_tag)
-    ch.stop_consuming()
 
 
 @when("the unaddressed batch is loaded, the print files are generated")
@@ -234,29 +214,6 @@ def check_message_redelivery_rate(context):
 def retrieve_hi_case(context):
     response = requests.get(f'{caseapi_url}{context.individual_case_id}')
     test_helper.assertEqual(response.status_code, 200, 'Case not found')
-
-
-def send_questionnaire_linked_msg_to_rabbit(questionnaire_id, case_id):
-    questionnaire_linked_message = {
-        'event': {
-            'type': 'QUESTIONNAIRE_LINKED',
-            'source': 'FIELDWORK_GATEWAY',
-            'channel': 'FIELD',
-            "dateTime": "2011-08-12T20:17:46.384Z",
-            "transactionId": "c45de4dc-3c3b-11e9-b210-d663bd873d93"
-        },
-        'payload': {
-            'uac': {
-                "caseId": case_id,
-                'questionnaireId': questionnaire_id,
-            }
-        }
-    }
-    with RabbitContext(exchange=Config.RABBITMQ_EVENT_EXCHANGE) as rabbit:
-        rabbit.publish_message(
-            message=json.dumps(questionnaire_linked_message),
-            content_type='application/json',
-            routing_key=Config.RABBITMQ_QUESTIONNAIRE_LINKED_ROUTING_KEY)
 
 
 def _send_individual_hh_questionnaire_linked_msg_to_rabbit(questionnaire_id, case_id, individual_case_id=None):
