+++ acceptance_tests/utilities/fulfilment_helper.py
@@ -1,0 +1,41 @@
+import json
+
+from acceptance_tests.utilities.rabbit_context import RabbitContext
+from config import Config
+
+
+def get_first_case(context):
+    return context.case_created_events[0]['payload']['collectionCase']
+
+
+def send_print_fulfilment_request(context, fulfilment_code):
+    context.first_case = get_first_case(context)
+
+    message = json.dumps(
+        {
+            "event": {
+                "type": "FULFILMENT_REQUESTED",
+                "source": "CONTACT_CENTRE_API",
+                "channel": "CC",
+                "dateTime": "2019-07-07T22:37:11.988+0000",
+                "transactionId": "d2541acb-230a-4ade-8123-eee2310c9143"
+            },
+            "payload": {
+                "fulfilmentRequest": {
+                    "fulfilmentCode": fulfilment_code,
+                    "caseId": context.first_case['id'],
+                    "contact": {
+                        "title": "Mrs",
+                        "forename": "Test",
+                        "surname": "McTest"
+                    }
+                }
+            }
+        }
+    )
+
+    with RabbitContext(exchange=Config.RABBITMQ_EVENT_EXCHANGE) as rabbit:
+        rabbit.publish_message(
+            message=message,
+            content_type='application/json',
+            routing_key=Config.RABBITMQ_FULFILMENT_REQUESTED_ROUTING_KEY)
